Add id-length rule
Checks the length of identifiers using ESLint and returns warnings for very short or very long IDs. This is not meant to fix hard to read variable names by itself or even meant to catch all names that are too short or too long. We're simply adding an automated check for the extreme cases.